+++ lib_test/ring_test.ml
@@ -14,6 +14,7 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *)
 
+[@@@warning "-27"]
 open OUnit
 
 let ( |> ) a b = b a